Issue-2121 move filter method to FileUtils
